Cast rlim_t to unsigned long

+++ cmdline.c
@@ -129,7 +129,7 @@ rlim_t cmdlineParseRLimit(int res, const char *optarg, unsigned long mul)
         LOG_F("RLIMIT %d needs a numeric or 'max'/'def' value ('%s' provided)", res, optarg);
     }
     rlim_t val = strtoul(optarg, NULL, 0) * mul;
-    if (val == ULONG_MAX && errno != 0) {
+    if ((unsigned long)val == ULONG_MAX && errno != 0) {
         PLOG_F("strtoul('%s', 0)", optarg);
     }
     return val;
